Improve comment support

+++ Korepetynder.Contracts/Responses/Tutors/TutorResponse.cs
@@ -4,8 +4,8 @@ namespace Korepetynder.Contracts.Responses.Tutors
     {
         public Guid UserId { get; set; }
         public IEnumerable<int> Locations { get; set; }
-        public int Score { get; set; }
-        public TutorResponse(Guid id, IEnumerable<int> locations, int score)
+        public decimal Score { get; set; }
+        public TutorResponse(Guid id, IEnumerable<int> locations, decimal score)
         {
             UserId = id;
             Locations = locations;
